+++ src/Operation/Internal/BaseOperation.php
@@ -33,9 +33,9 @@ abstract class BaseOperation implements Operation
         return $this->prev;
     }
     
-    public function streamingFinished(Signal $signal): void
+    public function streamingFinished(Signal $signal): bool
     {
-        $this->next->streamingFinished($signal);
+        return $this->next->streamingFinished($signal);
     }
     
     public function isLazy(): bool