+++ src/Operation/Scan.php
@@ -35,12 +35,12 @@ final class Scan extends BaseOperation
         $this->next->handle($signal);
     }
     
-    public function streamingFinished(Signal $signal): void
+    public function streamingFinished(Signal $signal): bool
     {
         //TODO check if it works properly in complex scenarios
         
         $this->handle($signal);
         
-        $this->next->streamingFinished($signal);
+        return $this->next->streamingFinished($signal);
     }
 }